# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agentic URL Shortener API
  version: 1.0.0
  description: URL shortener APIs plus governed engineering workflow demonstration endpoints.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/urls:
    post:
      summary: Create a shortened URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
                expiresInMinutes: { type: integer, minimum: 1 }
      responses:
        '200': { description: Short URL created }
        '400': { description: Invalid URL or expiration }
...
  /api/v1/urls/{code}/analytics:
    get:
      summary: Read click analytics
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      responses:
        '200': { description: Analytics returned }
  /api/v1/urls/{code}:
    delete:
      summary: Delete a short URL
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
...